fix a path-related bug on Windows
Under Windows, when Eclipse imports a project it will replace
'C:\foo\bar\baz' with 'foo/bar/baz', dropping the 'C:\' and
replacing backslashes with forward slashes.  This causes
a problem in Eclipse, because it can't resolve 'foo/bar/baz'
as being on the C drive.  This fix adds an explicit replaceAll
from backslashes to forward slashes, which prevents Eclipse
from doing it wrongly.