Made `createSrc` and `eclipseOutput` keys scoped properly.
When resolving `createSrc` and `eclipseOutput` keys we take into
account configuration so both of those keys can be scoped properly.

This allows one to have e.g.:

EclipseKeys.createSrc in Test := EclipseCreateSrc.ValueSet.empty

which makes sbteclipse to not include test directories in generated
.classpath file.

Updated a scripted test to show the effect of this change.