Multiple Scala version support for Scala IDE 4.0. Fix #239
The latest Scala IDE 4.0 supports both Scala 2.10 and 2.11 projects. To
correctly export a Scala 2.10 project, a few properties need to be properly set.

This commit assumes that sbteclipse only needs to support Scala IDE 4.0+, as I
haven't tested if the additional settings could cause issues when importing a
Scala 2.10 project in Scala IDE 3.x.

I've manually tested that running the `eclipse` command on both the sbteclipse
project itself, and the playframework project work as expected.

review by @huitseeker or  @dragos