Automatic detection of the project's flavor. Fix #277
When importing a stock Java project, with no dependency on Scala (usually,
because sbt setting `autoScalaLibrary` is set to `false`), it doesn't make sense
to add the Scala builder and nature to the Eclipse .classpath file.  In fact,
doing so it's actually going to cause issues as soon as the project is imported,
because the Scala nature is active but no scala-library.jar was found, and hence
the sbteclipse xml transformation for swapping the scala-library.jar with the
Scala bundle was not executed.

The proposed solution to fix the above problem is to have a new default flavour:
Autodetect. Autodetect will detect the right project's flavour based on the fact
that the scala-library.jar is, or isn't, in the classpath.

This commit, together with the fix for #239, makes the
`WithBundledScalaContainers` obsolete. First, the Scala bundle should always be
used when importing a Scala project, and it should never be used when it's a
Java project. The new Autodetect flavour takes care of this. Second, the actual
reason why `WithBundledScalaContainers` was introduced was to properly import
Scala 2.10 projects in the Scala IDE 4.0. #239 fixes exactly this problem (i.e.,
Scala 2.10 projects are now imported with the right scalac options), making
`WithBundledScalaContainers` indeed obsolete.