Allow overriding setting defaults at build-level
Fix #296.

Build settings can be set at the global level, at the build level or at
the project level, in order of increasing priority. To allow overriding
settings for a whole build, SBT guidelines recommend that the defaults
be set at the global level. This PR does that wherever possible.

In particular:
* Bump major version (because we'll break binary compatibility).
* Add globalEclipseSettings for Eclipse settings at the global
  priority (not build-level or project-level priority), and move
  most settings there.

  Exceptions:

  - `skipParents`, which is documented as a build-level setting.
    4bfd04bb7f241981041cba69865a104e8c34c6a6 duplicated the default to
    be at both build- and project-level, but the copy at the project
    level is redundant.
  - `skipProject`, which after #309 is set to true at the build level
    and cleared at the project level, except for projects that disable
    the plugin altogether.