Ensure Eclipse plugin works when disabled
In some situations, the eclipse plugin may be disabled on a particular
project.  If this happens, the eclipse command fails, saying the
skipProject setting can't be found for that project.  This ensures that
in that situation, skipProject defaults to true, so that it doesn't
fail, and the project is instead skipped.